+++ tests/test_chat_api.py
@@ -584,6 +584,7 @@ def test_get_journey_defaults(api_client: TestClient, logged_in_user):
         "use_anywhere": False,
         "onboarded": False,
         "successful_answers": 0,
+        "news_seen_version": 0,
     }
 
 
@@ -644,6 +645,7 @@ def test_put_journey_can_reset_explicit_false(api_client: TestClient, logged_in_
         "use_anywhere": False,
         "onboarded": False,
         "successful_answers": 0,
+        "news_seen_version": 0,
     }